// back/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const OrderStatus = {
  pending: 'pending',
  paid: 'paid',
  cancelled: 'cancelled'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const Role = {
  admin: 'admin',
  customer: 'customer'
} as const

export type Role = (typeof Role)[keyof typeof Role]